Predicates
Predicates have been added and managed.
The OptionMessageStates class allows the selection to be based off of player characteristics, the only one of which that is currently done is whether or not the player has completed a given quest, or default.
Also just changed some errors my ide told me.